Extract the resolution of the directory for a given namespace
This avoids duplicating the logic in the 2 writers (with slightly
different implementation) and will allow swapping the resolver for
standalone packages.
