+++ src/CodeGenerator/generate.php
@@ -5,6 +5,7 @@ use AsyncAws\CodeGenerator\File\Cache;
 use AsyncAws\CodeGenerator\File\CachedFileDumper;
 use AsyncAws\CodeGenerator\File\ClassWriter;
 use AsyncAws\CodeGenerator\File\ComposerWriter;
+use AsyncAws\CodeGenerator\File\Location\AsyncAwsMonoRepoResolver;
 use AsyncAws\CodeGenerator\Generator\ApiGenerator;
 use Symfony\Component\Console\Application;
 
@@ -14,7 +15,8 @@ $src = $_SERVER['ASYNC_AWS_GENERATE_SRC'] ?? __DIR__ . '/src';
 $cache = new Cache($_SERVER['ASYNC_AWS_GENERATE_CACHE'] ?? __DIR__ . '/.async-aws.cache');
 $manifest = $_SERVER['ASYNC_AWS_GENERATE_MANIFEST'] ?? __DIR__ . '/manifest.json';
 
-$command = new GenerateCommand($manifest, $cache, new ClassWriter($src, new CachedFileDumper($cache)), new ComposerWriter($src), new ApiGenerator());
+$directoryResolver = new AsyncAwsMonoRepoResolver($src);
+$command = new GenerateCommand($manifest, $cache, new ClassWriter($directoryResolver, new CachedFileDumper($cache)), new ComposerWriter($directoryResolver), new ApiGenerator());
 $application->add($command);
 $application->setDefaultCommand($command->getName(), true);
